Update FORIEGN KEYs to be BIGINT, since that's the actual type of SERIAL, not INT. Add support for a few more EMR tables.
git-svn-id: https://svn.master.freemedsoftware.org/freemed/trunk@2917 6507a214-7cff-0310-bd63-e3842a560bc8

+++ data/schema/mysql/allergies.sql
@@ -26,7 +26,7 @@ SOURCE data/schema/mysql/patient_emr.sql
 CREATE TABLE IF NOT EXISTS `allergies` (
 	allergy			VARCHAR (150) NOT NULL,
 	severity		VARCHAR (150) NOT NULL,
-	patient			INT UNSIGNED NOT NULL DEFAULT 0,
+	patient			BIGINT UNSIGNED NOT NULL DEFAULT 0,
 	reviewed		TIMESTAMP (14) NOT NULL DEFAULT NOW(),
 	id			SERIAL,
